refactor(ios): decompose IosGattServer 390->228 loc via extension functions
Extract lifecycle (open/close/delegate setup) to IosGattServerLifecycle.kt
and central tracking/sweep to IosGattServerCentrals.kt using the
established Approach B extension-function pattern.

Facade reduced from 390 to 228 lines (-42%), below the 300-line target.
All three files stay in the com.atruedev.kmpble.server package.

Visibility changes (private -> internal): manager, delegate,
centralIdleTimeout, centralSweepInterval, serviceDefinitions, isOpen,
isClosed, _connections, pendingServiceAdd.
